[BPK-4276] Restructure remaining D pages
