+++ docs/src/pages/DialogsPage/AndroidDialogPage.js
