+++ docs/src/pages/DialogsPage/DialogExamples.js
