+++ docs/src/pages/DialogsPage/DialogsPage.scss
